requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
emoji==2.14.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
socksio==1.0.0
yarl==1.20.1
